Refactor code / Fix notification error
