+++ src/handlers/errors/get.js
@@ -1,15 +1,17 @@
 'use strict';
 
 const console = require('console');
-const resgen = require('../../lib/resgen');
 const storage = require('../../lib/storage');
-const checkApiKey = require('../../lib/checkApiKey');
 const moment = require('moment');
 const {
     bucketName,
     errorByMessageTable,
     errorByTimeunitTable
 } = require('../../lib/constants');
+const {
+    resgen,
+    checkApiKey
+} = require('../../lib/functions');
 
 module.exports.list = (event, context, cb) => {
     // Check faultline API Key
