--- src/lib/timeunits.js
@@ -1,9 +1,0 @@
-'use strict';
-
-module.exports = {
-    year: 'YYYY',
-    month: 'YYYY-MM',
-    day: 'YYYY-MM-DD',
-    hour: 'YYYY-MM-DD HH',
-    minute: 'YYYY-MM-DD HH:mm'
-};
